requirements.txt
pony == 0.7.14
discord.py == 1.6.0
PyYAML == 5.3.1
pytz == 2020.5
google-api-python-client == 1.12.8
google-auth-httplib2 == 0.0.4
google-auth-oauthlib == 0.4.2